backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
Pillow==10.1.0
django-extensions==3.2.3
python-decouple==3.8
cryptography==41.0.7
requests==2.31.0
websockets==12.0
channels==4.0.0
channels-redis==4.1.0
djoser==2.2.0
djangorestframework-simplejwt==5.3.0
django-filter==23.4